web-service/requirements.txt
Django==3.2
mysqlclient==2.0.3
redis==3.5.3
python-dotenv==0.19.2
django-redis==4.12.1
djangorestframework==3.12.4
rq==1.10.0
djangorestframework-simplejwt==4.8.0
requests==2.32.3